Fix: double -> int

+++ main/src/main/java/main/main/laborcontract/entity/LaborContract.java
@@ -34,7 +34,7 @@ public class LaborContract {
 
     private LocalDateTime endOfContract;
 
-    private double basicSalary;
+    private int basicSalary;
 
     private LocalTime startTime;
